+++ src/pages/Home/components/CardsMenu/CardsMenu.test.tsx
@@ -5,16 +5,6 @@ import { getGeneratedPath } from 'shared/utils/path'
 import { withApp } from '../../../../../test/utils/withApp'
 
 describe('CardsMenu', () => {
-  it('has card with external link to Touringcar', async () => {
-    const pathToPage = getGeneratedPath(RouteIds.HOME)
-
-    withApp(pathToPage)
-
-    const cardWithExternalLink = screen.getByTestId('card-with-external-link')
-
-    expect(cardWithExternalLink).toHaveTextContent('Touringcar (Tour Buzz)')
-  })
-
   it('has card with src attribute', async () => {
     const pathToPage = getGeneratedPath(RouteIds.HOME)