+++ src/pages/Touringcar/components/MapLayers/MessagesLayer.tsx
@@ -2,13 +2,16 @@ import { useCallback, useContext, useEffect } from 'react'
 
 import { MapPanelContext, mapPanelConstants } from '@amsterdam/arm-core'
 import { useQuery } from '@tanstack/react-query'
-import getTouringcarMessages, { TouringcarMessage } from 'api/touringcar/messages'
+import getTouringcarMessages, { TouringcarMessage, TouringcarMessagePart } from 'api/touringcar/messages'
 import { format } from 'date-fns'
+import i18n from 'i18n'
 import { MapLayerId, MapPanelTab, useTouringcarMapContext } from 'pages/Touringcar/contexts/MapContext'
 import { MarkerClusterGroup } from 'shared/components/MapLayers/MarkerClusterGroup'
 import { DATE_FORMAT_REVERSED, DateHumanReadable_Year_Month_Day } from 'shared/utils/dateTime'
 
 import TouringcarMarker from '../Marker/Marker'
+
+type Language = 'nl' | 'en' | 'de' | 'es' | 'fr'
 
 export const MessagesLayer = () => {
   const { activeMapLayers, messagesDate, setCurrentMessage, setActiveTab } = useTouringcarMapContext()
@@ -27,25 +30,30 @@ export const MessagesLayer = () => {
 
   const findMessage = useCallback(
     (title: string) => {
-      const message = data?.features.find(item => item.properties?.title === title)
+      if (!data) return
+      const message: TouringcarMessage | undefined = data.features.find(
+        msg => msg?.properties[(i18n.language || 'nl') as Language].title === title
+      )
       setCurrentMessage(message)
     },
-    [data?.features, setCurrentMessage]
+    [data, setCurrentMessage]
   )
 
   const createClusterMarkers = () => {
-    return data!.features.map((item: TouringcarMessage) => {
-      const marker = TouringcarMarker(item, MapLayerId.touringcarMessagesLayerId)
+    return data!.features.map((message: TouringcarMessage) => {
+      const marker = TouringcarMarker(message, MapLayerId.touringcarMessagesLayerId)
 
-      let tooltipText = `<p><strong>${item.properties.title}</strong></p>
-      ${item.properties.body ?? `<p>${item.properties.body}</p>`}
-      ${item.properties.advice ?? `<p><strong>Advies:</strong> ${item.properties.advice}</p>`}`
+      const msgParts = message.properties[(i18n.language || 'nl') as Language]
+
+      let tooltipText = `<p><strong>${msgParts.title}</strong></p>
+      ${msgParts.body ?? `<p>${msgParts.body}</p>`}
+      ${msgParts.advice ?? `<p><strong>Advies:</strong> ${msgParts.advice}</p>`}`
 
       marker.bindTooltip(tooltipText)
 
       marker.on('click', () => {
         setActiveTab(MapPanelTab.MESSAGES)
-        findMessage(item.properties.title)
+        findMessage(msgParts.title)
         setPositionFromSnapPoint(mapPanelConstants.SnapPoint.Halfway)
       })